+++ studio/src/components/operations-overview.tsx
@@ -1,32 +1,23 @@
-import useWindowSize from "@/hooks/use-window-size";
-import { dateFormatter, useChartData } from "@/lib/insights-helpers";
-import { formatNumber } from "@/lib/utils";
-import { ExclamationTriangleIcon } from "@heroicons/react/24/outline";
-import { useQuery } from "@tanstack/react-query";
-import { getDashboardAnalyticsView } from "@wundergraph/cosmo-connect/dist/platform/v1/platform-PlatformService_connectquery";
+import useWindowSize from '@/hooks/use-window-size';
+import { dateFormatter, useChartData } from '@/lib/insights-helpers';
+import { formatNumber } from '@/lib/utils';
+import { ExclamationTriangleIcon } from '@heroicons/react/24/outline';
+import { useQuery } from '@tanstack/react-query';
 import {
-  OperationRequestCount,
-  RequestSeriesItem,
-} from "@wundergraph/cosmo-connect/dist/platform/v1/platform_pb";
-import { useId, useMemo } from "react";
-import {
-  Area,
-  AreaChart,
-  CartesianGrid,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from "recharts";
-import BarList from "./analytics/barlist";
-import { EmptyState } from "./empty-state";
-import { Badge } from "./ui/badge";
-import { Button } from "./ui/button";
-import { Loader } from "./ui/loader";
-import { Separator } from "./ui/separator";
-import { EnumStatusCode } from "@wundergraph/cosmo-connect/dist/common_pb";
-import { useRouter } from "next/router";
-import { constructAnalyticsTableQueryState } from "./analytics/constructAnalyticsTableQueryState";
+  getDashboardAnalyticsView,
+} from '@wundergraph/cosmo-connect/dist/platform/v1/platform-PlatformService_connectquery';
+import { OperationRequestCount, RequestSeriesItem } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
+import { useId, useMemo } from 'react';
+import { Area, AreaChart, CartesianGrid, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
+import BarList from './analytics/barlist';
+import { EmptyState } from './empty-state';
+import { Badge } from './ui/badge';
+import { Button } from './ui/button';
+import { Loader } from './ui/loader';
+import { Separator } from './ui/separator';
+import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common_pb';
+import { useRouter } from 'next/router';
+import { constructAnalyticsTableQueryState } from './analytics/constructAnalyticsTableQueryState';
 
 const valueFormatter = (number: number) => `${formatNumber(number)}`;
 
@@ -34,10 +25,10 @@ const CustomTooltip = ({ active, payload, label, valueLabel }: any) => {
   if (active && payload && payload.length) {
     return (
       <div className="rounded-md border border-popover bg-popover p-2 text-sm shadow-md outline-0 backdrop-blur-lg">
-        <p className="label">{label}</p>
-        <p className="intro">
+        <div className="label">{label}</div>
+        <div className="intro">
           {valueLabel || payload[0].name}: {payload[0].value}
-        </p>
+        </div>
       </div>
     );
   }