+++ cli/CHANGELOG.md
@@ -4,6 +4,22 @@ Images can be found [here](https://github.com/orgs/wundergraph/packages?repo_nam
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/wundergraph/cosmo/compare/wgc@0.6.5...wgc@0.7.0) (2023-09-02)
+
+### Features
+
+* add prometheus ([#31](https://github.com/wundergraph/cosmo/issues/31)) ([d318c73](https://github.com/wundergraph/cosmo/commit/d318c7331d77d21d0246344d76fbe0fc6b617174)) (@StarpTech)
+
+## [0.6.5](https://github.com/wundergraph/cosmo/compare/wgc@0.6.4...wgc@0.6.5) (2023-08-31)
+
+**Note:** Version bump only for package wgc
+
+## [0.6.4](https://github.com/wundergraph/cosmo/compare/wgc@0.6.3...wgc@0.6.4) (2023-08-31)
+
+### Bug Fixes
+
+* throw auth error inside try, allow to customize k8s probes, use … ([#24](https://github.com/wundergraph/cosmo/issues/24)) ([2d5695b](https://github.com/wundergraph/cosmo/commit/2d5695b95adad9b2fd8a6cacbc2dd2a1c2cb9bd6)) (@StarpTech)
+
 ## [0.6.3](https://github.com/wundergraph/cosmo/compare/wgc@0.6.2...wgc@0.6.3) (2023-08-29)
 
 **Note:** Version bump only for package wgc