+++ controlplane/CHANGELOG.md
@@ -4,6 +4,29 @@ Images can be found [here](https://github.com/orgs/wundergraph/packages?repo_nam
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.7.1...controlplane@0.8.0) (2023-09-02)
+
+### Features
+
+* add prometheus ([#31](https://github.com/wundergraph/cosmo/issues/31)) ([d318c73](https://github.com/wundergraph/cosmo/commit/d318c7331d77d21d0246344d76fbe0fc6b617174)) (@StarpTech)
+
+## [0.7.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.7.0...controlplane@0.7.1) (2023-08-31)
+
+### Bug Fixes
+
+* modify seed script ([#28](https://github.com/wundergraph/cosmo/issues/28)) ([7dba7ee](https://github.com/wundergraph/cosmo/commit/7dba7eed828648254be04d9f70585f2b930cd657)) (@JivusAyrus)
+
+# [0.7.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.6.2...controlplane@0.7.0) (2023-08-31)
+
+### Bug Fixes
+
+* analytics view and get trace query ([#23](https://github.com/wundergraph/cosmo/issues/23)) ([916488b](https://github.com/wundergraph/cosmo/commit/916488b697fa8f446e7aaa3d3f4b3e504671a23b)) (@JivusAyrus)
+* throw auth error inside try, allow to customize k8s probes, use … ([#24](https://github.com/wundergraph/cosmo/issues/24)) ([2d5695b](https://github.com/wundergraph/cosmo/commit/2d5695b95adad9b2fd8a6cacbc2dd2a1c2cb9bd6)) (@StarpTech)
+
+### Features
+
+* migrate graphs from apollo ([#17](https://github.com/wundergraph/cosmo/issues/17)) ([0d9d025](https://github.com/wundergraph/cosmo/commit/0d9d025adadf11fd0516cbe10f470765757a9853)) (@JivusAyrus)
+
 ## [0.6.2](https://github.com/wundergraph/cosmo/compare/controlplane@0.6.1...controlplane@0.6.2) (2023-08-29)
 
 **Note:** Version bump only for package controlplane